1、修改统一service接口返回类型 2、完成 商品->品牌管理模块接口 3、完成 商品->规格模板模块接口

+++ DogBro-Mall-Parent/DogBro-Mall-Service/Mall-Service-Goods/src/main/java/cn/byxll/service/impl/AlbumServiceImpl.java
@@ -5,6 +5,8 @@ import cn.byxll.goods.pojo.Album;
 import cn.byxll.service.AlbumService;
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
+import entity.Result;
+import entity.StatusCode;
 import org.apache.commons.lang.StringUtils;
 import org.springframework.stereotype.Service;
 import tk.mybatis.mapper.entity.Example;
@@ -26,88 +28,108 @@ public class AlbumServiceImpl implements AlbumService {
     /**
      * 新增相册
      * @param album 相册实体
+     * @return 响应数据
      */
     @Override
-    public void add(Album album) {
-        albumMapper.insert(album);
+    public Result<Boolean> add(Album album) {
+        if(album == null) { return new Result<>(false, StatusCode.ARGERROR, "参数异常"); }
+        int i = albumMapper.insert(album);
+        if(i>0) { return new Result<>(true, StatusCode.OK, "添加成功"); }
+        return new Result<>(false, StatusCode.ERROR, "操作失败");
     }
 
     /**
      * 删除相册
      * @param id 相册id
+     * @return  响应数据
      */
     @Override
-    public void delete(Long id) {
-        albumMapper.deleteByPrimaryKey(id);
+    public Result<Boolean> delete(Long id) {
+        if(id == null) { return new Result<>(false, StatusCode.ARGERROR, "参数异常"); }
+        int i = albumMapper.deleteByPrimaryKey(id);
+        if(i>0) { return new Result<>(true, StatusCode.OK, "删除成功"); }
+        return new Result<>(false, StatusCode.ERROR, "操作失败");
     }
 
     /**
      * 修改相册
-     * @param album 相册实体
+     * @param album 响应数据
      */
     @Override
-    public void update(Album album) {
-        albumMapper.updateByPrimaryKey(album);
+    public Result<Boolean> update(Album album) {
+        if(album == null) { return new Result<>(false, StatusCode.ARGERROR, "参数异常"); }
+        int i = albumMapper.updateByPrimaryKey(album);
+        if(i>0) { return new Result<>(true, StatusCode.OK, "操作成功"); }
+        return new Result<>(false, StatusCode.ERROR, "操作失败");
     }
 
     /**
      * 查询所有相册
-     * @return 相册实体集合
+     * @return 响应数据
      */
     @Override
-    public List<Album> findAll() {
-        return albumMapper.selectAll();
+    public Result<List<Album>> findAll() {
+        List<Album> albumList = albumMapper.selectAll();
+        return new Result<>(true, StatusCode.OK,"查询成功", albumList);
     }
 
     /**
      * 通过id查询相册
-     * @param id 相册id
-     * @return 相册实体
+     * @param id    相册id
+     * @return      响应数据
      */
     @Override
-    public Album findById(Long id) {
-        return albumMapper.selectByPrimaryKey(id);
+    public Result<Album> findById(Long id) {
+        if(id == null) { return new Result<>(false, StatusCode.ARGERROR, "参数异常",null); }
+        Album album = albumMapper.selectByPrimaryKey(id);
+        if(album == null) { return new Result<>(false, StatusCode.ERROR, "查询失败",null); }
+        return new Result<>(true, StatusCode.OK, "查询成功", album);
     }
 
     /**
      * 按条件查询相册集合
-     *
-     * @param album 相册条件实体
-     * @return 相册实体集合
+     * @param album     相册条件实体
+     * @return          响应数据
      */
     @Override
-    public List<Album> findListByParam(Album album) {
+    public Result<List<Album>> findListByParam(Album album) {
+        if(album == null) { return new Result<>(false, StatusCode.ARGERROR, "参数异常",null); }
         // 构建查询条件按
         Example example = createExample(album);
         // 通过构建的条件查询
-        return albumMapper.selectByExample(example);
+        List<Album> albumList = albumMapper.selectByExample(example);
+        return new Result<>(true, StatusCode.OK, "查询成功", albumList);
     }
 
     /**
      * 相册分页查询
-     * @param page     当前页码
-     * @param pageSize 每页大小
-     * @return 分页信息
+     * @param page      当前页码
+     * @param pageSize  每页大小
+     * @return          响应数据
      */
     @Override
-    public PageInfo<Album> findPage(Integer page, Integer pageSize) {
+    public Result<PageInfo<Album>> findPage(Integer page, Integer pageSize) {
+        if(page == null || pageSize == null) { return new Result<>(false, StatusCode.ARGERROR, "参数异常",null); }
         // 静态分页
         PageHelper.startPage(page,pageSize);
-        return new PageInfo<>(albumMapper.selectAll());
+        return new Result<>(true, StatusCode.OK, "查询成功", new PageInfo<>(albumMapper.selectAll()));
     }
 
     /**
      * 根据条件分页查询
-     * @param album    相册条件实体
-     * @param page     当前页码
-     * @param pageSize 每页大小
-     * @return 分页信息
+     * @param album     相册条件实体
+     * @param page      当前页码
+     * @param pageSize  每页大小
+     * @return          响应数据
      */
     @Override
-    public PageInfo<Album> findPageByParam(Album album, Integer page, Integer pageSize) {
+    public Result<PageInfo<Album>> findPageByParam(Album album, Integer page, Integer pageSize) {
+        if(album == null || page == null || pageSize == null) {
+            return new Result<>(false, StatusCode.ARGERROR, "参数异常");
+        }
         Example example = createExample(album);
         PageHelper.startPage(page,pageSize);
-        return new PageInfo<>(albumMapper.selectByExample(example));
+        return new Result<>(true, StatusCode.OK, "查询成功", new PageInfo<>(albumMapper.selectByExample(example)));
     }
 
     /**
